# fix: math: optimize and test FormatInt + simplify LegacyNewDecFromStr