--- a/math/CHANGELOG.md
+++ b/math/CHANGELOG.md
@@ -31,5 +31,6 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+* [#14010](https://github.com/cosmos/cosmos-sdk/pull/14010) Optimize FormatInt to not do plain string concentation when formatting thousands and instead build it more efficiently.
 * [#13381](https://github.com/cosmos/cosmos-sdk/pull/13381) Add uint `IsNil` method.
 * [#12634](https://github.com/cosmos/cosmos-sdk/pull/12634) Move `sdk.Dec` to math package, call it `LegacyDec`.