@odeke-em 2022-11-25
> Id like to stay away from such optimisations that complicates code, if we don't see this actually being a performance issue. 
> 
> 
> 
> Could you provide a profile or something similar that shows this on a mainnet node? 

Hey @tac0turtle, this is for value renderers. I got here by auditing the new code that has recently been added (it's has been requested that Orijtech continues to do feature based audits every quarter which you too have asked for; this code is going to run on ledger devices,) but I don't see this "complicating code" in fact this further simplifies the code to read more naturally how one adds thousands separators compared to the prior form.

I got here by looking through benchmarks in tx/valuerenderer and their respective profiles and noticing that DecimalValueRenderer code consumed lots of unnecessary cycles in string concatenation.