@tac0turtle 2022-11-25
Oh, if this isn't run in the state machine then I don't think we need this., I want to leave this up to @amaurym since he is leading the project.