@odeke-em 2022-11-29
Thanks @amaurym @tac0turtle @julienrbrt for the reviews. I've made the update to the change log.